feat move hardware menu in system tab
