+++ main.py
@@ -198,18 +198,6 @@ class MainUi(PyQt6.QtWidgets.QMainWindow, base_ui.WidgetUI, base_ui.Communicatio
 
         self.actionActive_threads.triggered.connect(self.active_threads_dlg.show)
 
-        self.actionRestore_chip_config.triggered.connect(self.load_flashdump_from_file)
-        self.serialchooser.connected.connect(self.actionRestore_chip_config.setEnabled)
-
-        self.actionSave_chip_config.triggered.connect(self.save_flashdump_to_file)
-        self.serialchooser.connected.connect(self.actionSave_chip_config.setEnabled)
-
-        self.actionReboot.triggered.connect(self.reboot)
-        self.serialchooser.connected.connect(self.actionReboot.setEnabled)
-
-        self.actionReset_Factory_Config.triggered.connect(self.reset_factory_btn)
-        self.serialchooser.connected.connect(self.actionReset_Factory_Config.setEnabled)
-
         self.actionEffectsMonitor.triggered.connect(self.effects_monitor_dlg.display)
         #self.serialchooser.connected.connect(self.actionEffectsMonitor.setEnabled)
 
@@ -291,12 +279,6 @@ class MainUi(PyQt6.QtWidgets.QMainWindow, base_ui.WidgetUI, base_ui.Communicatio
             self.menuLanguage.addAction(action)
             action.toggled.connect(self.change_lang_callback)
 
-
-    def reboot(self):
-        """Send the reboot message to the board."""
-        self.send_command("sys", "reboot")
-        self.reconnect()
-
     def check_configurator_update(self):
         """Checks if there is an update for the configurator only"""
         donotnotify = self.profile_ui.get_global_setting("donotnotify_updates",False)
@@ -363,37 +345,6 @@ class MainUi(PyQt6.QtWidgets.QMainWindow, base_ui.WidgetUI, base_ui.Communicatio
         self.send_value("sys","debug",1 if enabled else 0)
         # Reload mainclasses
         self.serialchooser.get_main_classes() # TODO better move somewhere else
-
-
-    def save_flashdump_to_file(self):
-        """Send a async message to get the flashdump from board."""
-        self.get_value_async("sys", "flashdump", config.saveDump)
-
-    def load_flashdump_from_file(self):
-        """Load dumpfile and send config to board."""
-        dump = config.loadDump()
-        if not dump:
-            return
-
-        if self.connected:
-            for sector in dump["flash"]:
-                self.send_value("sys", "flashraw", sector["val"], sector["addr"], 0)
-            # Message
-            msg = PyQt6.QtWidgets.QMessageBox(
-                PyQt6.QtWidgets.QMessageBox.Icon.Information,
-                self.tr("Restore flash dump"),
-                self.tr("Uploaded flash dump.\nPlease reboot."),
-            )
-        else:
-            # Message
-            msg = PyQt6.QtWidgets.QMessageBox(
-                PyQt6.QtWidgets.QMessageBox.Icon.Warning,
-                self.tr("Can't restore flash dump"),
-                self.tr("Please connect board first."),
-            )
-
-
-        msg.exec()
 
     def timeout_check_cb(self, port_checked):
         """Close the serial connection if the port is not open after a timeout."""
@@ -838,26 +789,6 @@ class MainUi(PyQt6.QtWidgets.QMainWindow, base_ui.WidgetUI, base_ui.Communicatio
             self.log("Disconnected")
             self.reset_tabs()
             self.comms.removeAllCallbacks() # Ensure everything is cleared
-
-    def reset_factory(self, btn):
-        """Send a async message to reset factory settings."""
-        cmd = btn.text()
-        if cmd == "OK":
-            self.send_value("sys", "format", 1)
-            self.send_command("sys", "reboot")
-            self.reset_port()
-
-    def reset_factory_btn(self):
-        """Prompt a confirmation to the user when he click on reset factory."""
-        msg = PyQt6.QtWidgets.QMessageBox()
-        msg.setIcon(PyQt6.QtWidgets.QMessageBox.Icon.Warning)
-        msg.setText(self.tr("Format flash and reset?"))
-        msg.setStandardButtons(
-            PyQt6.QtWidgets.QMessageBox.StandardButton.Ok
-            | PyQt6.QtWidgets.QMessageBox.StandardButton.Cancel
-        )
-        msg.buttonClicked.connect(self.reset_factory) # pylint: disable=no-value-for-parameter
-        msg.exec()
 
     def change_profile(self, profilename: str):
         """Change the current profile by this one selected."""
