+++ serial_ui.py
@@ -44,9 +44,6 @@ class SerialChooser(base_ui.WidgetUI, base_ui.CommunicationHandler):
 
         self.pushButton_refresh.clicked.connect(self.get_ports)
         self.pushButton_connect.clicked.connect(self.serial_connect_button)
-        #self.pushButton_send.clicked.connect(self.send_line)
-        #self.lineEdit_cmd.returnPressed.connect(self.send_line)
-        self.pushButton_mainclasschange.clicked.connect(self.main_btn)
         
         self.update()
 
@@ -66,23 +63,22 @@ class SerialChooser(base_ui.WidgetUI, base_ui.CommunicationHandler):
         """
         self.hidden.emit()
 
-    def serial_log(self, txt):
-        """Add a new text in the history widget."""
-        if isinstance(txt, list):
-            txt = "\n".join(txt)
-        else:
-            txt = str(txt)
-        #self.serialLogBox.append(txt)
+    # def serial_log(self, txt):
+    #     """Add a new text in the history widget."""
+    #     if isinstance(txt, list):
+    #         txt = "\n".join(txt)
+    #     else:
+    #         txt = str(txt)
 
-    def send_line(self):
-        """Read the command input text, display it in history widget and send it to the board."""
-        cmd = self.lineEdit_cmd.text() + "\n"
-        self.serial_log(">" + cmd)
-        self.serial_write_raw(cmd)
+    # def send_line(self):
+    #     """Read the command input text, display it in history widget and send it to the board."""
+    #     cmd = self.lineEdit_cmd.text() + "\n"
+    #     self.serial_log(">" + cmd)
+    #     self.serial_write_raw(cmd)
 
-    def write(self, data):
-        """Write data to the serial port."""
-        self._serial.write(data)
+    # def write(self, data):
+    #     """Write data to the serial port."""
+    #     self._serial.write(data)
 
     def update(self):
         """Update the UI when a connection is successfull.
@@ -94,27 +90,12 @@ class SerialChooser(base_ui.WidgetUI, base_ui.CommunicationHandler):
             self.pushButton_connect.setText(self.tr("Disconnect"))
             self.comboBox_port.setEnabled(False)
             self.pushButton_refresh.setVisible(False)
-            
-            self.label.setVisible(False)
-            self.comboBox_main.setVisible(False)
-            self.pushButton_mainclasschange.setVisible(False)
-                                          
-            #self.pushButton_send.setEnabled(True)
-            #self.lineEdit_cmd.setEnabled(True)
             self.connected.emit(True)
-            self.get_main_classes()
         else:
             self.pushButton_connect.setText(self.tr("Connect"))
             self.comboBox_port.setEnabled(True)
             self.pushButton_refresh.setVisible(True)
-            
-            self.label.setVisible(False)
-            self.comboBox_main.setVisible(False)
-            self.pushButton_mainclasschange.setVisible(False)
-            #self.pushButton_send.setEnabled(False)
-            #self.lineEdit_cmd.setEnabled(False)
             self.connected.emit(False)
-            #self.groupBox_system.setEnabled(False)
 
     def serial_connect_button(self):
         """Check if it's not connected, and call start the serial connection."""
@@ -208,46 +189,3 @@ class SerialChooser(base_ui.WidgetUI, base_ui.CommunicationHandler):
     def auto_connect(self, nb_compatible_device):
         if (nb_compatible_device == 1) :
             self.serial_connect_button()
-
-    def update_mains(self, dat):
-        """Parse the list of main classes received from board, and update the combobox."""
-        self.comboBox_main.clear()
-        self._class_ids, self._classes = helper.classlistToIds(dat)
-
-        if self.main_id is None:
-            # self.main.resetPort()
-            self.groupBox_system.setEnabled(False)
-            return
-        #self.groupBox_system.setEnabled(True)
-
-        helper.updateClassComboBox(
-            self.comboBox_main, self._class_ids, self._classes, self.main_id
-        )
-
-        self.main.log("Detected mode: " + self.comboBox_main.currentText())
-        self.main.update_tabs()
-
-    def get_main_classes(self):
-        """Get the main classes available from the board in Async."""
-
-        def fct(i):
-            """Store the main currently selected to refresh the UI."""
-            self.main_id = i
-
-        self.get_value_async("main", "id", fct, conversion=int, delete=True)
-        self.get_value_async("sys", "lsmain", self.update_mains, delete=True)
-
-    def main_btn(self):
-        """Read the select main class in the combobox.
-
-        Push it to the board and display the reload warning.
-        """
-        index = self._classes[self.comboBox_main.currentIndex()][0]
-        self.send_value("sys", "main", index)
-        self.main.reconnect()
-        msg = PyQt6.QtWidgets.QMessageBox(
-            PyQt6.QtWidgets.QMessageBox.Icon.Information,
-            "Main class changed",
-            "Chip is rebooting. Please reconnect.",
-        )
-        msg.exec()
